allunbot/requirements.txt
Django==3.2.18
gunicorn==20.1.0
whitenoise==6.2.0
Brotli==1.0.9
pymongo[tls,srv]==3.7.2  # mongo db
python-dotenv==0.10.1
pyTelegramBotAPI==4.10.0
pyTelegramBotAPI-markups==0.0.4
requests==2.28.2
jsonschema==4.17.3
beautifulsoup4==4.11.2